# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat API Documentation",
        "description": "This is the API documentation for the Chat application",
        "contact": {
            "email": "support@yourapp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/chats/{chat_id}/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Retrieve messages for a chat",
                "description": "Fetch all messages associated with a specific chat.",
                "operationId": "28b2cf0854a1d25abff9c4cd9d367d57",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Hello World!"
                                            },
                                            "sender": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "sent"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send a message in a chat",
                "description": "Create a new message in a specific chat.",
                "operationId": "b18ccc803be91a78bbb91b9d038bc62f",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, how are you?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Hello World!"
                                            },
                                            "sender": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "sent"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/chats/{chat_id}/messages/{message_id}": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Delete a message from a chat",
                "description": "Remove a specific message from a chat.",
                "operationId": "2a485a37644173ea4b9acc324398bc1d",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "This message has been deleted."
                                            },
                                            "sender": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "deleted"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat or message not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Messages"
                ],
                "summary": "Update a message in a chat",
                "description": "Modify an existing message in a specific chat.",
                "operationId": "593d4f6271637c8ea0226fc42b83f3cd",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "message_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message_text"
                                ],
                                "properties": {
                                    "message_text": {
                                        "type": "string",
                                        "example": "Hello World! (Edited)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Hello World! (Edited)"
                                            },
                                            "sender": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "edited"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat or message not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}